# BackEnd/docs/swagger.yaml
basePath: /
definitions:
...
  dto.TransactionRequest:
    properties:
      account_id:
        example: acc_123456789
        type: string
      amount:
        example: 125.5
        type: number
      budget_id:
        example: budget_555666777
        type: string
      category_id:
        example: cat_987654321
        type: string
      description:
        example: Weekly grocery shopping at Walmart
        type: string
      name:
        example: Grocery Shopping
        type: string
      type_transation:
        enum:
        - income
        - expense
        example: expense
        type: string
    required:
    - amount
    - name
    type: object
  dto.TransactionResponse:
    properties:
      account_id:
        type: string
      amount:
        type: number
      budget_id:
        type: string
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      type_transation:
        type: string
    required:
    - amount
    - name
    type: object
host: localhost:8080
info:
  contact:
    email: support@gestordePresupuesto.com
    name: API Support
    url: http://github.com/osmait/gestorDePresupuesto
  description: API for managing personal budget, transactions, accounts, and investments
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Budget Management System API
  version: "1.0"
paths:
...
  /transaction:
    get:
      consumes:
      - application/json
      description: Retrieve all transactions for the authenticated user across all
        their accounts
      parameters:
      - description: Start date (YYYY/MM/DD)
        in: query
        name: date
        type: string
      - description: End date (YYYY/MM/DD)
        in: query
        name: date2
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            items:
              $ref: '#/definitions/dto.TransactionResponse'
            type: array
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Get all transactions across all accounts
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Create a new financial transaction (income or expense)
      parameters:
      - description: Transaction creation data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/dto.TransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Create a new transaction
      tags:
      - Transactions
  /transaction/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a specific transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Delete a transaction
      tags:
      - Transactions
    get:
      consumes:
      - application/json
      description: Retrieve transactions for a specific account within a date range
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: string
      - description: Start date (YYYY/MM/DD)
        in: query
        name: date
        type: string
      - description: End date (YYYY/MM/DD)
        in: query
        name: date2
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions for the account
          schema:
            items:
              $ref: '#/definitions/dto.TransactionResponse'
            type: array
        "401":
          description: Unauthorized - Invalid JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Account not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Get transactions for a specific account
      tags:
      - Transactions
securityDefinitions:
  JWT:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"