# BackEnd/internal/platform/server/handler/transaction/transactionHandler.go
package transaction

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	dto "github.com/osmait/gestorDePresupuesto/internal/platform/dto/transaction"
	errorHandler "github.com/osmait/gestorDePresupuesto/internal/platform/server/handler/error"
	"github.com/osmait/gestorDePresupuesto/internal/services/transaction"
	"github.com/rs/zerolog/log"
)

// CreateTransaction godoc
//
//	@Summary		Create a new transaction
//	@Description	Create a new financial transaction (income or expense)
//	@Tags			Transactions
//	@Accept			json
//	@Produce		json
//	@Security		JWT
//	@Param			transaction	body		dto.TransactionRequest	true	"Transaction creation data"
//	@Success		201			{object}	map[string]string		"Transaction created successfully"
//	@Failure		400			{object}	map[string]string		"Bad request - Invalid input"
//	@Failure		401			{object}	map[string]string		"Unauthorized - Invalid JWT token"
//	@Failure		500			{object}	map[string]string		"Internal server error"
//	@Router			/transaction [post]
func CreateTransaction(transactionservice *transaction.TransactionService) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		userId := ctx.GetString("X-User-Id")

		var transactionRequest dto.TransactionRequest

		if err := ctx.BindJSON(&transactionRequest); err != nil {
			ctx.JSON(http.StatusBadRequest, "Error fields required ")
			return
		}
		err := transactionservice.CreateTransaction(
			ctx,
			transactionRequest.Name,
			transactionRequest.Description,
			transactionRequest.Amount,
			transactionRequest.TypeTransation,
			transactionRequest.AccountId,
			userId,
			transactionRequest.CategoryId,
			transactionRequest.BudgetId,
		)
		if err != nil {
			errorHandler.ReponseByTypeOfErr(err, ctx)
			return
		}

		ctx.JSON(http.StatusCreated, "Transaction created")
	}
}

// FindAllTransactionOfAllAccount godoc
//
//	@Summary		Get all transactions across all accounts
//	@Description	Retrieve all transactions for the authenticated user across all their accounts
//	@Tags			Transactions
//	@Accept			json
//	@Produce		json
//	@Security		JWT
//	@Param			date	query		string					false	"Start date (YYYY/MM/DD)"
//	@Param			date2	query		string					false	"End date (YYYY/MM/DD)"
//	@Success		200		{array}		dto.TransactionResponse	"List of transactions"
//	@Failure		401		{object}	map[string]string		"Unauthorized - Invalid JWT token"
//	@Failure		500		{object}	map[string]string		"Internal server error"
//	@Router			/transaction [get]
func FindAllTransactionOfAllAccount(transactionService *transaction.TransactionService) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		date1 := ctx.Query("date")
		date2 := ctx.Query("date2")
		userId := ctx.GetString("X-User-Id")
		if date1 == "" || date2 == "" {
			currentTime := time.Now()
			// Use proper date arithmetic to avoid negative days
			date1Time := currentTime.AddDate(0, 0, -7) // 7 days ago
			date2Time := currentTime.AddDate(0, 0, 2)  // 1 day from now

			// Format dates properly
			date1 = fmt.Sprintf("%d/%d/%d", date1Time.Year(), date1Time.Month(), date1Time.Day())
			date2 = fmt.Sprintf("%d/%d/%d", date2Time.Year(), date2Time.Month(), date2Time.Day())
		}

		transactionsList, err := transactionService.FindAllOfAllAccounts(ctx, userId)
		if err != nil {
			errorHandler.ReponseByTypeOfErr(err, ctx)
			return
		}
		log.Debug().Int("total_transactions", len(transactionsList)).Msg("retrieved all transactions")

		// Count income transactions for debug
		incomeCount := 0
		for _, t := range transactionsList {
			if t.TypeTransation == "income" {
				incomeCount++
			}
		}
		log.Debug().Int("income_transactions", incomeCount).Msg("income transactions found")

		ctx.JSON(http.StatusOK, transactionsList)
	}
}

// FindAllTransaction godoc
//
//	@Summary		Get transactions for a specific account
//	@Description	Retrieve transactions for a specific account within a date range
//	@Tags			Transactions
//	@Accept			json
//	@Produce		json
//	@Security		JWT
//	@Param			id		path		string					true	"Account ID"
//	@Param			date	query		string					false	"Start date (YYYY/MM/DD)"
//	@Param			date2	query		string					false	"End date (YYYY/MM/DD)"
//	@Success		200		{array}		dto.TransactionResponse	"List of transactions for the account"
//	@Failure		401		{object}	map[string]string		"Unauthorized - Invalid JWT token"
//	@Failure		404		{object}	map[string]string		"Account not found"
//	@Failure		500		{object}	map[string]string		"Internal server error"
//	@Router			/transaction/{id} [get]
func FindAllTransaction(transactionService *transaction.TransactionService) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		date1 := ctx.Query("date")
		date2 := ctx.Query("date2")
		id := ctx.Param("id")

		if date1 == "" || date2 == "" {
			currentTime := time.Now()
			// Use proper date arithmetic to avoid negative days
			date1Time := currentTime.AddDate(0, 0, -2) // 2 days ago
			date2Time := currentTime.AddDate(0, 0, 1)  // 1 day from now

			// Format dates properly
			date1 = fmt.Sprintf("%d/%d/%d", date1Time.Year(), date1Time.Month(), date1Time.Day())
			date2 = fmt.Sprintf("%d/%d/%d", date2Time.Year(), date2Time.Month(), date2Time.Day())
		}

		transactionsList, err := transactionService.FindAll(ctx, date1, date2, id)
		if err != nil {
			errorHandler.ReponseByTypeOfErr(err, ctx)
			return
		}

		ctx.JSON(http.StatusOK, transactionsList)
	}
}

// DeleteTransaction godoc
//
//	@Summary		Delete a transaction
//	@Description	Delete a specific transaction by ID
//	@Tags			Transactions
//	@Accept			json
//	@Produce		json
//	@Security		JWT
//	@Param			id	path		string				true	"Transaction ID"
//	@Success		200	{object}	map[string]string	"Transaction deleted successfully"
//	@Failure		401	{object}	map[string]string	"Unauthorized - Invalid JWT token"
//	@Failure		404	{object}	map[string]string	"Transaction not found"
//	@Failure		500	{object}	map[string]string	"Internal server error"
//	@Router			/transaction/{id} [delete]
func DeleteTransaction(transactionService *transaction.TransactionService) gin.HandlerFunc {
	return func(ctx *gin.Context) {
		id := ctx.Param("id")
		err := transactionService.DeleteTransaction(ctx, id)
		if err != nil {
			errorHandler.ReponseByTypeOfErr(err, ctx)
			return
		}
		ctx.JSON(http.StatusOK, "Deleted")
	}
}